flyprint/pyproject.toml
[project]
name = "flyprint"
version = "0.1.0"
description = "Local print agent for FlyPush label printing"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Giorgio Gilestro", email = "giorgio@gilest.ro"}
]
keywords = ["printing", "labels", "cups", "flypush", "drosophila"]

dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
cups = [
    "pycups>=2.0.0",
]
gui = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]
windows = [
    "pywin32>=306",
]
build = [
    "pyinstaller>=6.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
flyprint = "flyprint.cli:main"

[project.gui-scripts]
flyprint-gui = "flyprint.app_entry:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flyprint"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B"]
